+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Removed `IndexMap` dependency
 - Reduce allocations when writing and parsing tags file
+- Reduce allocations in code path for creating `Tag` struct by native parsers
+  with extension fields support
 
 - [0.12.0]